+++ src/go/rpk/pkg/cli/cmd/redpanda.go
@@ -28,6 +28,7 @@ func NewRedpandaCommand(
 	command.AddCommand(redpanda.NewStartCommand(fs, mgr, launcher))
 	command.AddCommand(redpanda.NewStopCommand(fs, mgr))
 	command.AddCommand(redpanda.NewCheckCommand(fs, mgr))
+	command.AddCommand(redpanda.NewTuneCommand(fs, mgr))
 
 	return command
 }
